requirements.txt
pandas
requests
bs4
gscholar
bibtexparser
pyecharts==1.9.1
pyperclip
PyYAML